Move share/view dialog to the dialog folder (Dialog)
