+++ Atarashii/src/net/somethingdreadful/MAL/ProfileActivity.java
@@ -1,8 +1,6 @@
 package net.somethingdreadful.MAL;
 
-import android.app.AlertDialog;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
@@ -11,10 +9,10 @@ import android.support.v7.app.ActionBar;
 import android.support.v7.app.ActionBarActivity;
 import android.widget.Toast;
 
-import net.somethingdreadful.MAL.account.AccountService;
 import net.somethingdreadful.MAL.adapters.ProfilePagerAdapter;
 import net.somethingdreadful.MAL.api.MALApi;
 import net.somethingdreadful.MAL.api.response.User;
+import net.somethingdreadful.MAL.dialog.ShareDialogFragment;
 import net.somethingdreadful.MAL.tasks.UserNetworkTask;
 import net.somethingdreadful.MAL.tasks.UserNetworkTaskFinishedListener;
 
@@ -87,10 +85,10 @@ public class ProfileActivity extends ActionBarActivity implements UserNetworkTas
                     startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse("http://myanimelist.net/profile/" + record.getName())));
                 break;
             case R.id.View:
-                choosedialog(false);
+                showShareDialog(false);
                 break;
             case R.id.Share:
-                choosedialog(true);
+                showShareDialog(true);
         }
         return true;
     }
@@ -108,54 +106,13 @@ public class ProfileActivity extends ActionBarActivity implements UserNetworkTas
         }
     }
 
-    void choosedialog(final boolean share) { //as the name says
-        AlertDialog.Builder builder = new AlertDialog.Builder(this);
-        final Intent sharingIntent = new Intent(android.content.Intent.ACTION_SEND);
-        if (share) {
-            builder.setTitle(R.string.dialog_title_share);
-            builder.setMessage(R.string.dialog_message_share);
-            sharingIntent.setType("text/plain");
-            sharingIntent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_WHEN_TASK_RESET);
-        } else {
-            builder.setTitle(R.string.dialog_title_view);
-            builder.setMessage(R.string.dialog_message_view);
-        }
-
-        builder.setPositiveButton(R.string.dialog_label_animelist, new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                if (share) {
-                    sharingIntent.putExtra(Intent.EXTRA_TEXT, getResources().getString(R.string.share_animelist)
-                            .replace("$name;", record.getName())
-                            .replace("$username;", AccountService.getUsername()));
-                    startActivity(Intent.createChooser(sharingIntent, getString(R.string.dialog_title_share_via)));
-                } else {
-                    Uri mallisturlanime = Uri.parse("http://myanimelist.net/animelist/" + record.getName());
-                    startActivity(new Intent(Intent.ACTION_VIEW, mallisturlanime));
-                }
-            }
-        });
-        builder.setNeutralButton(R.string.dialog_label_cancel, new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-
-            }
-        });
-        builder.setNegativeButton(R.string.dialog_label_mangalist, new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                if (share) {
-                    sharingIntent.putExtra(Intent.EXTRA_TEXT, getResources().getString(R.string.share_mangalist)
-                            .replace("$name;", record.getName())
-                            .replace("$username;", AccountService.getUsername()));
-                    startActivity(Intent.createChooser(sharingIntent, getString(R.string.dialog_title_share_via)));
-                } else {
-                    Uri mallisturlmanga = Uri.parse("http://myanimelist.net/mangalist/" + record.getName());
-                    startActivity(new Intent(Intent.ACTION_VIEW, mallisturlmanga));
-                }
-            }
-        });
-        builder.show();
+    private void showShareDialog(boolean type){
+        ShareDialogFragment share = new ShareDialogFragment();
+        Bundle args = new Bundle();
+        args.putString("title", record.getName());
+        args.putBoolean("share", type);
+        share.setArguments(args);
+        share.show(getFragmentManager(), "fragment_share");
     }
 
     @Override
